build: do not hard-code `AM_DEFAULT_VERBOSITY` in Makefile.iptrules
Currently it is set to zero, which means that the default in
Makefile.iptrules is not consistent with that in the other Makefiles,
and passing `--disable-silent-rules` to configure cannot be used to
change it.

Set it to `@AM_DEFAULT_VERBOSITY@` instead, which will be expanded to the
appropriate default value.

Signed-off-by: Jeremy Sowden <jeremy@azazel.net>