build: fix inclusion of Makefile.extra
Commit 08a16d90ceae ("build: use `$(top_srcdir)` when including Makefile.extra")
replaces hard-coded relative paths used to include Makefile.extra with
variables.  However, despite the commit message, the variables are enclosed with
braces, not parentheses, and it transpires that automake does not support the
use of braces in this context.  As a result, Makefile.extra is not included, and
the libxt_ACCOUNT.so and libxt_pknock.so extensions are not built.

Use parentheses instead.

Link: https://bugs.launchpad.net/ubuntu/+source/xtables-addons/+bug/2080528
Fixes: 08a16d90ceae ("build: use `$(top_srcdir)` when including Makefile.extra")
Signed-off-by: Jeremy Sowden <jeremy@azazel.net>